pyproject.toml
[project]
name = "macos-clipboard-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "applescript",
    "build>=1.2.2.post1",
    "mcp[cli]>=1.7.1",
    "pillow>=11.2.1",
]

[project.scripts]
macos-clipboard-mcp = "main:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
Pillow = "^10.0.0"
